fix: custom hook & rename
