+++ package.json
@@ -31,7 +31,6 @@
     "@reduxjs/toolkit": "^1.8.2",
     "aws-sdk": "^2.1231.0",
     "axios": "^0.27.2",
-    "dotenv": "^16.0.3",
     "framer-motion": "^6",
     "html2canvas": "^1.4.1",
     "jspdf": "^2.5.1",
