+++ src/__tests__/redux/canvasControlSlice.test.ts
@@ -39,13 +39,13 @@ it("should reset all states and scales to initial", () => {
     dragActivate: true,
     dragStart: true,
     resetState: false,
-    screenshot: null,
+    courtDataUrl: null,
   };
   expect(reducer(previousState, resetAll())).toEqual({
     zoomScale: 1,
     dragStart: false,
     dragActivate: false,
     resetState: true,
-    screenshot: null,
+    courtDataUrl: null,
   });
 });
