+++ src/components/HalfCourt/index.tsx
@@ -1,6 +1,6 @@
 import { Stage, Layer, Group } from "react-konva";
-import { Flex, useToast } from "@chakra-ui/react";
-import { ReactReduxContext, Provider, useDispatch } from "react-redux";
+import { Flex } from "@chakra-ui/react";
+import { ReactReduxContext, Provider } from "react-redux";
 import ThreePointArea from "../BasketballCourt/ThreePointArea";
 import KeyArea from "../BasketballCourt/KeyArea";
 import CourtArea from "../BasketballCourt/CourtArea";
@@ -13,14 +13,10 @@ import useCourt from "@/hooks/useCourt";
 import { IZoomShift } from "@/interfaces/zoomShift";
 import { useRef, useEffect } from "react";
 import canvasControlModel from "../../utils/canvasControlModel";
-import { useStoreSelector } from "@/store/hooks";
-import { updateCourtStage } from "@/utils/uploadImage";
+import useImageDataUrl from "@/hooks/useImageDataUrl";
 
 const HalfCourt = () => {
-  const dispatch = useDispatch();
-  const toast = useToast();
   const { courtAreaXLength, courtAreaYLength, borderLength, court, courtStartPoint } = useCourt();
-  const rulerState = useStoreSelector((state) => state.buttonToggle.isRulerOn);
   const ref = useRef<any>(null);
 
   const zoomShift: IZoomShift = {
@@ -41,9 +37,7 @@ const HalfCourt = () => {
     ref.current.y(0);
   }, [canvasContorl.canvasStates.resetState]);
 
-  useEffect(() => {
-    updateCourtStage(dispatch, ref, rulerState, toast);
-  }, [canvasStates.selectedColor, borderLength]);
+  useImageDataUrl(ref);
 
   return (
     <Flex
